live-voting-nextjs/requirements.txt
python-dotenv==1.0.1
supabase==2.5.1
vecs==0.4.3
psycopg2-binary==2.9.9